+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import lotto.exception.DuplicateLottoNumberException;
 import lotto.exception.InvalidCountOfLottoNumberException;
