+++ src/main/java/lotto/domain/lotto/LottoRepository.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import lotto.domain.lottogenerator.LottoGeneratingStrategy;
 import lotto.domain.lottogenerator.LottoGenerator;
