+++ src/main/java/lotto/domain/lotto/Rank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import java.util.Arrays;
 import java.util.Map;
