+++ src/main/java/lotto/domain/lottogenerator/RandomLottoGeneratingStrategy.java
@@ -1,7 +1,7 @@
 package lotto.domain.lottogenerator;
 
-import lotto.domain.Lotto;
-import lotto.domain.LottoNumber;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 
 import java.util.Collections;
 import java.util.List;
