+++ src/main/java/lotto/domain/paymentinfo/Payment.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.paymentinfo;
 
 import lotto.exception.PaymentOutOfBoundsException;
