+++ src/test/java/lotto/domain/lotto/LottoTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 import lotto.exception.DuplicateLottoNumberException;
 import lotto.exception.InvalidCountOfLottoNumberException;
 import org.junit.jupiter.api.Test;
