+++ src/test/java/lotto/domain/lotto/LottoTicketsTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.lotto.LottoRepository;
+import lotto.domain.lotto.LottoTickets;
 import lotto.domain.lottogenerator.ManualLottoGeneratingStrategy;
 import org.junit.jupiter.api.Test;
