+++ src/test/java/lotto/domain/lotto/RankTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
+import lotto.domain.lotto.Rank;
 import org.junit.jupiter.api.Test;
 
 import java.util.HashMap;
