+++ src/test/java/lotto/domain/lotto/WinningLottoTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import lotto.domain.lottogenerator.ManualLottoGeneratingStrategy;
 import lotto.exception.DuplicateLottoNumberException;
