+++ src/test/java/lotto/domain/lottogenerator/LottoGeneratorTest.java
@@ -1,7 +1,7 @@
 package lotto.domain.lottogenerator;
 
-import lotto.domain.Lotto;
-import lotto.domain.LottoNumber;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
