+++ src/test/java/lotto/domain/paymentinfo/CountOfLottoTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.paymentinfo;
 
+import lotto.domain.paymentinfo.CountOfLotto;
+import lotto.domain.paymentinfo.Payment;
 import lotto.exception.NaturalNumberException;
 import lotto.exception.PaymentOutOfBoundsException;
 import org.junit.jupiter.api.BeforeEach;
