+++ src/test/java/lotto/domain/paymentinfo/PaymentTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.domain.paymentinfo;
 
+import lotto.domain.paymentinfo.Payment;
 import lotto.exception.PaymentOutOfBoundsException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
